+++ Cxx11/nstream-occa.cc
@@ -61,6 +61,8 @@
 ///
 //////////////////////////////////////////////////////////////////////
 
+#include "occa.hpp"
+
 #include "prk_util.h"
 
 int main(int argc, char * argv[])
