+++ Cxx11/p2p-hyperplane-sycl.cc
@@ -59,6 +59,8 @@
 ///
 //////////////////////////////////////////////////////////////////////
 
+#include "CL/sycl.hpp"
+
 #include "prk_util.h"
 #include "p2p-kernel.h"
