+++ Cxx11/prk_tbb.h
@@ -1,0 +1,54 @@
+///
+/// Copyright (c) 2018, Intel Corporation
+///
+/// Redistribution and use in source and binary forms, with or without
+/// modification, are permitted provided that the following conditions
+/// are met:
+///
+/// * Redistributions of source code must retain the above copyright
+///       notice, this list of conditions and the following disclaimer.
+/// * Redistributions in binary form must reproduce the above
+///       copyright notice, this list of conditions and the following
+///       disclaimer in the documentation and/or other materials provided
+///       with the distribution.
+/// * Neither the name of Intel Corporation nor the names of its
+///       contributors may be used to endorse or promote products
+///       derived from this software without specific prior written
+///       permission.
+///
+/// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+/// "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+/// LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
+/// FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
+/// COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+/// INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
+/// BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+/// LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+/// CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+/// LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+/// ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+/// POSSIBILITY OF SUCH DAMAGE.
+
+#ifndef PRK_TBB_H
+#define PRK_TBB_H
+
+#ifdef USE_TBB
+# include <tbb/tbb.h>
+# include <tbb/parallel_for.h>
+# include <tbb/blocked_range.h>
+# if ( PRK_TBB_PARTITIONER == 1)
+//#  warning STATIC
+   tbb::static_partitioner tbb_partitioner;
+# elif ( PRK_TBB_PARTITIONER == 2)
+//#  warning AFFINITY
+   tbb::affinity_partitioner tbb_partitioner;
+# elif ( PRK_TBB_PARTITIONER == 3)
+//#  warning SIMPLE
+   tbb::simple_partitioner tbb_partitioner;
+# else
+//#  warning AUTO
+   tbb::auto_partitioner tbb_partitioner;
+# endif
+#endif
+
+#endif /* PRK_TBB_H */
