+++ Cxx11/prk_util.h
@@ -64,22 +64,6 @@
 #include <numeric>
 #include <algorithm>
 
-template<class I, class T>
-const T prk_reduce(I first, I last, T init) {
-#if (defined(__cplusplus) && (__cplusplus >= 201703L)) && !defined(__GNUC__)
-    return std::reduce(first, last, init);
-#elif (defined(__cplusplus) && (__cplusplus >= 201103L))
-    return std::accumulate(first, last, init);
-#else
-    // unreachable, but preserved as reference implementation
-    T r(0);
-    for (I i=first; i!=last; ++i) {
-        r += *i;
-    }
-    return r;
-#endif
-}
-
 // These headers are busted with NVCC and GCC 5.4.0
 // The <future> header is busted with Cray C++ 8.6.1.
 #if !defined(__NVCC__) && !defined(_CRAYC)
@@ -87,174 +71,59 @@ const T prk_reduce(I first, I last, T init) {
 #include <future>
 #endif
 
-#define PRAGMA(x) _Pragma(#x)
+#include "prk_simd.h"
 
-#ifdef _OPENMP
-# include <omp.h>
-# define OMP(x) PRAGMA(omp x)
-# define OMP_PARALLEL(x) PRAGMA(omp parallel x)
-# define OMP_PARALLEL_FOR_REDUCE(x) PRAGMA(omp parallel for reduction (x) )
-# define OMP_MASTER PRAGMA(omp master)
-# define OMP_BARRIER PRAGMA(omp barrier)
-# define OMP_FOR(x) PRAGMA(omp for x)
-# define OMP_FOR_REDUCE(x) PRAGMA(omp for reduction (x) )
-// OpenMP SIMD if supported, else not.
-# if (_OPENMP >= 201300)
-#  define OMP_SIMD PRAGMA(omp simd)
-#  define OMP_FOR_SIMD PRAGMA(omp for simd)
-#  define OMP_TASK(x) PRAGMA(omp task x)
-#  define OMP_TASKLOOP(x) PRAGMA(omp taskloop x )
-#  if defined(__INTEL_COMPILER)
-#   define OMP_TASKLOOP_COLLAPSE(n,x) PRAGMA(omp taskloop x )
-#  else
-#   define OMP_TASKLOOP_COLLAPSE(n,x) PRAGMA(omp taskloop collapse(n) x )
-#  endif
-#  define OMP_TASKWAIT PRAGMA(omp taskwait)
-#  define OMP_ORDERED(x) PRAGMA(omp ordered x)
-#  define OMP_TARGET(x) PRAGMA(omp target x)
-#  define OMP_DECLARE_TARGET PRAGMA(omp declare target)
-#  define OMP_END_DECLARE_TARGET PRAGMA(omp end declare target)
-# else
-#  define OMP_SIMD
-#  define OMP_FOR_SIMD PRAGMA(omp for)
-#  define OMP_TASK(x)
-#  define OMP_TASKLOOP(x)
-#  define OMP_TASKLOOP_COLLAPSE(n,x)
-#  define OMP_TASKWAIT
-#  define OMP_ORDERED(x)
-#  define OMP_TARGET(x)
-#  define OMP_DECLARE_TARGET
-#  define OMP_END_DECLARE_TARGET
-# endif
-#else
-# define OMP(x)
-# define OMP_PARALLEL(x)
-# define OMP_PARALLEL_FOR_REDUCE(x)
-# define OMP_MASTER
-# define OMP_BARRIER
-# define OMP_FOR(x)
-# define OMP_FOR_REDUCE(x)
-# define OMP_SIMD
-# define OMP_FOR_SIMD
-# define OMP_TASK(x)
-# define OMP_TASKLOOP(x)
-# define OMP_TASKLOOP_COLLAPSE(n,x)
-# define OMP_TASKWAIT
-# define OMP_ORDERED(x)
-# define OMP_TARGET(x)
-# define OMP_DECLARE_TARGET
-# define OMP_END_DECLARE_TARGET
+#ifdef USE_RANGES
+# include "prk_ranges.h"
 #endif
 
-#if defined(__INTEL_COMPILER)
-# define PRAGMA_SIMD PRAGMA(vector) PRAGMA(ivdep)
-// According to https://github.com/LLNL/RAJA/pull/310, this improves lambda performance
-# define PRAGMA_INLINE PRAGMA(forceinline recursive)
-#elif defined(__GNUC__) && defined(__GNUC_MINOR__) && ( ( (__GNUC__ == 4) && (__GNUC_MINOR__ == 9) ) || (__GNUC__ >= 5) )
-# define PRAGMA_SIMD PRAGMA(GCC ivdep)
-# define PRAGMA_INLINE PRAGMA(inline)
-#elif defined(__clang__)
-# define PRAGMA_SIMD PRAGMA(clang loop vectorize(assume_safety))
-# define PRAGMA_INLINE
-#else
-# define PRAGMA_SIMD
-# define PRAGMA_INLINE
+#ifdef USE_OPENMP
+# include "prk_openmp.h"
 #endif
 
 #ifdef USE_TBB
-# include <tbb/tbb.h>
-# include <tbb/parallel_for.h>
-# include <tbb/blocked_range.h>
-# if ( PRK_TBB_PARTITIONER == 1)
-//#  warning STATIC
-   tbb::static_partitioner tbb_partitioner;
-# elif ( PRK_TBB_PARTITIONER == 2)
-//#  warning AFFINITY
-   tbb::affinity_partitioner tbb_partitioner;
-# elif ( PRK_TBB_PARTITIONER == 3)
-//#  warning SIMPLE
-   tbb::simple_partitioner tbb_partitioner;
-# else
-//#  warning AUTO
-   tbb::auto_partitioner tbb_partitioner;
-# endif
-#endif
-
-#if defined(USE_RANGES)
-# if defined(USE_BOOST_IRANGE)
-#  include "boost/range/irange.hpp"
-# elif defined(USE_RANGES_TS)
-#  include "range/v3/view/iota.hpp"
-#  include "range/v3/view/slice.hpp"
-#  include "range/v3/view/stride.hpp"
-# else
-#  error You have not provided a version of ranges to use.
-# endif
-#endif
-
-#if defined(USE_BOOST_COMPUTE)
-# include "boost/compute.hpp"
-# include "boost/compute/container/valarray.hpp"
-#endif
-
-#if defined(__INTEL_COMPILER) && (__INTEL_COMPILER >= 1800)
-#define USE_INTEL_PSTL
+# include "prk_tbb.h"
 #endif
 
 #ifdef USE_PSTL
-# ifdef USE_INTEL_PSTL
-#  include <pstl/execution>
-#  include <pstl/algorithm>
-#  include <pstl/numeric>
-#  include <pstl/memory>
-# elif defined(__GNUC__) && defined(__GNUC_MINOR__) && \
-       ( (__GNUC__ >= 8) || (__GNUC__ == 7) && (__GNUC_MINOR__ >= 2) )
-#  include <parallel/algorithm>
-#  include <parallel/numeric>
-# endif
+# include "prk_pstl.h"
 #endif
 
 #ifdef USE_KOKKOS
-# include <Kokkos_Core.hpp>
-# include <Kokkos_Concepts.hpp>
-# include <Kokkos_MemoryTraits.hpp>
+# include "prk_kokkos.h"
 #endif
 
 #ifdef USE_RAJA
-# define RAJA_ENABLE_NESTED 1
-# include "RAJA/RAJA.hpp"
+# include "prk_raja.h"
 #endif
 
 #ifdef USE_THRUST
-# ifdef __NVCC__
-#  include <thrust/device_vector.h>
-# endif
-# include <thrust/host_vector.h>
-# include <thrust/fill.h>
-# include <thrust/sequence.h>
-# include <thrust/for_each.h>
-# include <thrust/transform.h>
-# include <thrust/transform_reduce.h>
-# include <thrust/iterator/counting_iterator.h>
-# include <thrust/execution_policy.h>
-# include <thrust/functional.h>
-#endif
-
-#ifdef USE_SYCL
-# include "CL/sycl.hpp"
-#endif
-
-#ifdef USE_OCCA
-# include "occa.hpp"
+# include "prk_thrust.h"
 #endif
 
 #define RESTRICT __restrict__
 
 namespace prk {
 
+    template<class I, class T>
+    const T reduce(I first, I last, T init) {
+#if (defined(__cplusplus) && (__cplusplus >= 201703L)) && !defined(__GNUC__)
+        return std::reduce(first, last, init);
+#elif (defined(__cplusplus) && (__cplusplus >= 201103L))
+        return std::accumulate(first, last, init);
+#else
+        // unreachable, but preserved as reference implementation
+        T r(0);
+        for (I i=first; i!=last; ++i) {
+            r += *i;
+        }
+        return r;
+#endif
+    }
+
     static inline double wtime(void)
     {
-#ifdef _OPENMP
+#if defined(USE_OPENMP) && defined(_OPENMP)
         return omp_get_wtime();
 #else
         using t = std::chrono::high_resolution_clock;
@@ -269,29 +138,6 @@ namespace prk {
     template <class T1, class T2>
     static inline auto divceil(T1 numerator, T2 denominator) -> decltype(numerator / denominator) {
         return ( numerator / denominator + (numerator % denominator > 0) );
-    }
-
-    template <class S, class E>
-    auto range(S start, E end) {
-#if defined(USE_BOOST_IRANGE)
-        return boost::irange(static_cast<decltype(end)>(start), end);
-#elif defined(USE_RANGES_TS)
-        return ranges::view::iota(static_cast<decltype(end)>(start), end);
-#endif
-    }
-
-    template <class S, class E, class B>
-    auto range(S start, E end, B blocking) {
-#if defined(USE_BOOST_IRANGE)
-        return boost::irange(static_cast<decltype(end)>(start), end, static_cast<decltype(end)>(blocking) );
-#elif defined(USE_RANGES_TS)
-        // NOTE:
-        // iota(s) | slice(s,e) | stride(b)  is faster than
-        // iota(s,e) | stride(b) for some reason.
-        return ranges::view::iota(static_cast<decltype(end)>(start)) |
-               ranges::view::slice(static_cast<decltype(end)>(start), end) |
-               ranges::view::stride(static_cast<decltype(end)>(blocking));
-#endif
     }
 
 } // namespace prk
