+++ Cxx11/stencil-sycl.cc
@@ -60,6 +60,8 @@
 ///
 //////////////////////////////////////////////////////////////////////
 
+#include "CL/sycl.hpp"
+
 #include "prk_util.h"
 #include "stencil_sycl.hpp"
