+++ Cxx11/transpose-occa.cc
@@ -51,6 +51,8 @@
 ///
 //////////////////////////////////////////////////////////////////////
 
+#include "occa.hpp"
+
 #include "prk_util.h"
 
 int main(int argc, char * argv[])
