+++ Cxx11/transpose-sycl.cc
@@ -49,6 +49,8 @@
 ///
 //////////////////////////////////////////////////////////////////////
 
+#include "CL/sycl.hpp"
+
 #include "prk_util.h"
 
 int main(int argc, char * argv[])
